rewrite/code_puppy-main-elixir-first-rewrite/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codepp"
version = "0.0.455"
description = "Code generation agent"
readme = "README.md"
requires-python = ">=3.14"
# Minimum version constraints. Exact versions are pinned in uv.lock.
# CI MUST use `uv sync --frozen` to enforce the lockfile.
# To update: `uv lock --upgrade-package <pkg>` then commit uv.lock.
# Dependencies are pinned for reproducibility. Review for security updates monthly.
dependencies = [
    "pydantic-ai-slim[openai,anthropic,mcp]==1.60.0",
    "typer>=0.24.1",
    "mcp>=1.26.0",
    "httpx[http2]>=0.28.1",
    "requests>=2.33.0",
    "rich>=14.3.3",
    "prompt_toolkit>=3.0.50",
    "textual>=8.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.3",
    "json-repair>=0.58.7",
    "pyfiglet>=1.0.4",
    "openai>=2.30.0",
    "ripgrep==15.0.0",
    "tenacity>=9.1.4",
    "playwright>=1.58.0",
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
    "Pillow>=12.1.1",
    "anthropic==0.79.0",
    "pyjwt>=2.0,<3.0",
    "pathspec>=0.12.0",
]
authors = [
    {name = "Michael Pfaffenberger"}
]
license = {text = "MIT"}
keywords = ["ai", "agent", "coding", "llm", "cli", "mcp", "terminal"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Code Generators",
    "Environment :: Console",
    "Intended Audience :: Developers",
]

[project.urls]
repository = "https://github.com/mpfaffenberger/code_puppy"
HomePage = "https://github.com/mpfaffenberger/code_puppy"

[project.optional-dependencies]
rust = ["maturin>=1.12.6"]
dbos = [
    "dbos>=2.16.0",
    "psycopg[c]>=3.3.0",
]

[project.scripts]
code-puppy = "code_puppy.main:main_entry"
pup = "code_puppy.main:main_entry"
gac = "code_puppy.plugins.git_auto_commit.cli:main"

[tool.logfire]
ignore_no_config = true

[tool.hatch.build]
packages = ["code_puppy"]

[tool.hatch.build.targets.wheel]
# Force-include Rust source directories into the wheel
# These will be placed at the wheel root alongside code_puppy/
[tool.hatch.build.targets.wheel.force-include]
"Cargo.toml" = "Cargo.toml"
"Cargo.lock" = "Cargo.lock"
"code_puppy_core/Cargo.toml" = "code_puppy_core/Cargo.toml"
"code_puppy_core/pyproject.toml" = "code_puppy_core/pyproject.toml"
"code_puppy_core/src" = "code_puppy_core/src"
"turbo_parse/Cargo.toml" = "turbo_parse/Cargo.toml"
"turbo_parse/pyproject.toml" = "turbo_parse/pyproject.toml"
"turbo_parse/src" = "turbo_parse/src"
"turbo_parse/queries" = "turbo_parse/queries"

[tool.ruff]
target-version = "py314"

[tool.ruff.lint.per-file-ignores]
"code_puppy/main.py" = ["E402"]
"code_puppy/cli_runner.py" = ["E402"]
"code_puppy/messaging/__init__.py" = ["E402"]
"code_puppy/config.py" = ["F821"]  # Lazy-evaluated path constants via __getattr__

[tool.pytest.ini_options]
addopts = "-n auto --dist=loadgroup --cov=code_puppy --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "eval: LLM evaluation test (skipped unless RUN_EVALS=1 is set)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial: marks tests that must not run in parallel with other tests",
]

[tool.coverage.run]
# Enable parallel mode to prevent race conditions when running with pytest-xdist
parallel = true
# omit = ["code_puppy/main.py"]  # Temporarily disabled for complete coverage tracking

[tool.coverage.report]
fail_under = 70
# Per-module coverage gates for plugins (audit-driven quality)
# See docs/CONTRIBUTING.md for coverage maintenance guidelines
show_missing = true
skip_covered = false

# Plugin modules requiring >=85% coverage
# Format: pattern = minimum percentage
[tool.coverage.plugins]
paths = [
    "code_puppy/plugins/pack_parallelism/*",
]
min_coverage = 85

[tool.uv]
python-preference = "only-managed"
override-dependencies = [
    "psycopg[c]>=3.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.6.0",
    "ruff>=0.15.8",
    "pexpect>=4.9.0",
    "maturin>=1.12.6",
    "hypothesis>=6.150.0",
    "pytest-randomly>=4.0.1",
]
